+++ src/smt/bitvec/mod.rs
@@ -13,13 +13,13 @@ pub enum BVType {
 pub type SymByte = BitVec<8>;
 
 #[derive(Debug, Clone, Eq)]
-pub struct BitVec<const SZ: u32> {
+pub struct BitVec<const SZ: usize> {
     pub inner: BVType,
     pub(crate) typ: super::SolverType,
 }
 
 
-// impl<const SZ: u32> PartialEq for BitVec<SZ> {
+// impl<const SZ: usize> PartialEq for BitVec<SZ> {
 //     fn eq(&self, other: &Self) -> bool {
 //         let BVType::Z3(a) = &self.inner;
 //         let BVType::Z3(b) = &other.inner;
@@ -28,7 +28,7 @@ pub struct BitVec<const SZ: u32> {
 //     }
 // }
 
-impl<const SZ: u32> Hash for BitVec<SZ> {
+impl<const SZ: usize> Hash for BitVec<SZ> {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.as_ref().hash(state);
         state.finish();
@@ -38,7 +38,7 @@ impl<const SZ: u32> Hash for BitVec<SZ> {
 }
 
 
-impl<const SZ: u32> BitVec<SZ> {
+impl<const SZ: usize> BitVec<SZ> {
     pub fn with_bv(bv: BV<'static>) -> Self {
         Self {
             inner: BVType::Z3(bv),
@@ -50,17 +50,17 @@ impl<const SZ: u32> BitVec<SZ> {
         self.inner = BVType::Z3(bv.simplify());
     }
 }
-impl<const SZ: u32> Default for BitVec<SZ> {
+impl<const SZ: usize> Default for BitVec<SZ> {
     fn default() -> Self {
         let ctx = ctx();
         Self {
-            inner: BVType::Z3(BV::from_u64(ctx, 0, SZ * 8)),
+            inner: BVType::Z3(BV::from_u64(ctx, 0, (SZ * 8) as u32)),
             typ: SolverType::Z3,
         }
     }
 }
 
-impl<const SZ: u32> PartialEq for BitVec<SZ> {
+impl<const SZ: usize> PartialEq for BitVec<SZ> {
     fn eq(&self, other: &Self) -> bool {
 
         let a = self.as_ref();
